requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
streamlit>=1.28.0
plotly>=5.17.0
numpy>=1.24.0
